handle serial buffer overflow condition (while the modem protocol *can* send frames >512 bytes the serial port handler can only handle frames up to 512 bytes); better handle transmitting double length (512 byte) P25 PDU frames;

+++ p25/P25TX.h
@@ -65,7 +65,7 @@ namespace p25
          * @param length Length of buffer.
          * @returns uint8_t Reason code.
          */
-        uint8_t writeData(const uint8_t* data, uint8_t length);
+        uint8_t writeData(const uint8_t* data, uint16_t length);
 
         /**
          * @brief Clears the local buffer.
